requirements.txt
alembic==1.0.8
aniso8601==3.0.2
arrow==0.13.1
Babel==2.6.0
beautifulsoup4==4.7.1
blinker==1.4
Bootstrap-Flask==1.1.0
certifi==2019.6.16
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-Babel==0.12.2
Flask-Cors==3.0.7
Flask-GraphQL==2.0.0
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.4.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
graphene==2.1.3
graphene-sqlalchemy==2.1.1
graphql-core==2.1
graphql-relay==0.4.5
graphql-server-core==1.1.1
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.0.8
MarkupSafe==1.1.1
Pillow==6.0.0
promise==2.2.1
psycopg2-binary==2.8.2
PyJWT==1.7.1
python-dateutil==2.8.0
python-dotenv==0.10.1
python-editor==1.0.4
python-slugify==3.0.2
pytz==2018.9
requests==2.22.0
Rx==1.6.1
singledispatch==3.4.0.3
six==1.12.0
soupsieve==1.9.1
SQLAlchemy==1.3.2
text-unidecode==1.2
urllib3==1.25.3
Werkzeug==0.15.2
WTForms==2.2.1
